Use cleveref for references.
Cleveref gives us two benefits.  First, it requires less LaTeX code for
references.  Instead of

  Section~\ref{foo}

we only need to write

  \Cref{foo}

Besides, cleveref allows us to make both the resulting label and the
number clickable and not just the number.  That makes the PDF easier to
navigate.  Hyperref's \autoref also does this but it doesn't handle
capitalisation correctly.